backend/requirements.txt
dnspython==2.4.2
APScheduler==3.10.4
requests==2.31.0
pytz==2025.2
fastapi==0.110.0
uvicorn==0.28.0
jinja2==3.1.3
